% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grademploy.R
\name{employee}
\alias{employee}
\title{Employee graph}
\usage{
employee(major = NULL)
}
\arguments{
\item{major}{The name of major in the USA}
}
\description{
This function illustrates a graph of number of jobs that require employee degree.
}
