% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grademploy.R
\name{max}
\alias{max}
\alias{min}
\title{Maximum and Minimum}
\usage{
max(variable = NULL)

min(variable = NULL)
}
\arguments{
\item{variable}{is the numeric variables in university dataset.}
}
\description{
These functions illustrate Maximum and Minimum.
}
