% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-min.R
\name{maxvar}
\alias{maxvar}
\alias{minvar}
\title{Maximum and Minimum}
\usage{
maxvar(variable = NULL)

minvar(variable = NULL)
}
\arguments{
\item{variable}{is the numeric variables in university dataset.}
}
\description{
These functions illustrate Maximum and Minimum.
}
