% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankgraduate-package.R
\docType{package}
\name{rankgraduate-package}
\alias{rankgraduate}
\alias{rankgraduate-package}
\title{rankgraduate: What the Package Does (One Line, Title Case)}
\description{
Provides datasets and functions to show graph of Shiny app that illustrates the rank of major in the USA by median income.
}
\author{
\strong{Maintainer}: Aphiaut Imuan \email{aimu0001@student.monash.edu}

}
\keyword{internal}
