% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-rank.R
\name{table_rank}
\alias{table_rank}
\title{table}
\usage{
table_rank(maxIncome = NULL, minIncome = NULL)
}
\arguments{
\item{maxIncome}{is the max median income that users set it.}

\item{minIncome}{is the min median income that users set it.}
}
\description{
This function illustrate five numbers
}
