% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grademploy.R
\name{employee}
\alias{employee}
\title{Employee graph}
\usage{
employee(major = NULL)
}
\arguments{
\item{major}{is the name of majors in the USA}
}
\description{
This function illustrates a graph of the number of jobs that require an employee degree.
}
