% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankgraduate-package.R
\docType{package}
\name{rankgraduate-package}
\alias{rankgraduate}
\alias{rankgraduate-package}
\title{rankgraduate: Rank Graduated Package}
\description{
Provides datasets and functions to show graph of Shiny app that illustrates the rank of major in the USA by median income.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rankgraduate.netlify.app/}
}

}
\author{
\strong{Maintainer}: Aphiaut Imuan \email{aimu0001@student.monash.edu}

}
\keyword{internal}
